requirements.txt
Flask
torch
pillow
torchvision
tensorboard
scikit-learn